backend/requirements.txt
Django==4.1.13
django-cors-headers==4.3.1
django-graphql-jwt==0.4.0
djangorestframework==3.14.0
djongo==1.3.6
graphene==3.4.3
graphene-django==3.2.2
PyJWT==2.12.1
pymongo==3.12.3
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
sqlparse==0.2.4
